ChatBackend/src/core/security.py
import jwt
import datetime
from core.config import SECRET_KEY

def generate_token(username: str):
    payload = {
        "sub": username,
        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=1),
    }
    return jwt.encode(payload, SECRET_KEY, algorithm="HS256")
