ChatBackend/src/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==2.10.0
python-jose==3.3.0
pyjwt==2.8.0
